+++ src/components/assets/EventListItem/EventListItem.jsx
@@ -1,0 +1,48 @@
+import React from 'react';
+import {Col} from 'react-onsenui';
+import moduleClass from './EventListItem.module.css';
+
+class EventListItem extends React.Component {
+  render() {
+    return (
+      <Col width={"360px"}>
+        <div className={moduleClass.eventCard} onClick={this.props.onClick()}>
+          {(()=> {
+            if (this.props.event.hasOwnProperty('photoURL')) {
+              return (
+                <div
+                  className={moduleClass.eventCardImage}
+                >
+                  <img
+                    className={moduleClass.eventCardImage}
+                    src={this.props.event.photoURL}
+                    alt="event"/>
+                </div>
+              );
+            } else {
+              return (
+                <div className={moduleClass.eventCardImage}>
+                  <div
+                    className={moduleClass.eventCardNoImage}
+                  >
+                    <div
+                      style={{
+                        padding: '5px',
+                        backgroundColor: 'rgba(0, 0, 0, .12)',
+                        boxShadow: '0 3px 2px rgba(0, 0, 0, .12)'}}
+                    >no Image</div>
+                  </div>
+                </div>
+              );
+            }
+          })()}
+          <div className={moduleClass.eventCardText}>
+            {this.props.event.name}
+          </div>
+        </div>
+      </Col>
+    );
+  }
+}
+
+export default EventListItem;
