+++ src/components/assets/EventListItem/EventListItem.module.css
@@ -1,0 +1,38 @@
+.eventCard {
+    font-family: -apple-system, 'Helvetica Neue', 'Helvetica', 'Arial', 'Lucida Grande', sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    font-weight: 400;
+    box-shadow: 0 1px 2px rgba(0, 0, 0, .12);
+    border-radius: 8px;
+    background-color: white;
+    box-sizing: border-box;
+    display: block;
+    margin: 8px;
+    text-align: left;
+    word-wrap: break-word;
+}
+
+.eventCardImage {
+    border-radius: 8px;
+    width: 344px;
+    height: 172px;
+    object-fit: cover;
+    box-shadow: 0 3px 2px rgba(0, 0, 0, .12);
+}
+.eventCardNoImage {
+    border-radius: 8px;
+    display: flex;
+    height: 100%;
+    background-color: #b3b3b3;
+    justify-content: center;
+    align-items: center;
+    object-fit: cover;
+ }
+
+.eventCardText {
+    height: 32px;
+    display: flex;
+    align-items: center;
+    padding: 0 8px !important;
+}
